+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## 0.5.0
+- Added the function to set a random wallpaper from favorite settings at start up.
+
 ## 0.4.0
 - Added the function to unregister from favorite wallpaper settings
 - Added the function to random playback of slide